split out platform building now that there are 4
